// lens-src/types/core.ts
export interface DocumentPosition {
  doc_id: string;
  file_path: string;
  line: number;
  col: number;
  length: number;
}